% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epierror.R
\docType{package}
\name{epierror}
\alias{epierror}
\alias{epierror-package}
\title{epierror: Quotes for the Epidemiolgist with an Error}
\description{
Easily include moody quotes in your error messages or emails.
}
\author{
\strong{Maintainer}: Philip Crain \email{philip.crain@doh.wa.gov}

}
