% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getquotes.R
\name{getquotes}
\alias{getquotes}
\title{Select some random quotes}
\usage{
getquotes(n)
}
\arguments{
\item{n}{Integer or Numeric element: Supplies the number of quotes/rows to be returned. If \code{n} is larger than the number of rows in \code{quotations}, then the entire df is returned in a random order.}
}
\value{
data.frame with randomly selected rows of quote, author, and book.
}
\description{
Select \code{n} number of quotes from \code{quotations} data.frame (included in package) and return the quotes, authors, and book titles as a data.frame.
}
\examples{
getquotes(1)
getquotes(5)['book']
getquotes(999)$author

}
\seealso{
\itemize{
\item \link[epierror]{selectquote} selects quote and returns as a single character element
\item \link[epierror]{selectquote.html} selects quote and returns as a single character element with HTML formatting
}
}
\author{
Philip Crain \href{mailto:philip.crain@doh.wa.gov}{philip.crain@doh.wa.gov}
}
