% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotations.R
\docType{data}
\name{quotations}
\alias{quotations}
\title{Book Quotes about Errors for Epidemiologists}
\format{
A data frame of 26 rows and 3 columns
\describe{
\item{quote}{Character values with edited book quotes}
\item{author}{Character values giving the original quotation's author}
\item{book}{Character values giving the source of the original quotation}
}
}
\source{
Many books and the creative minds of Data Integration and Quality Assurance
}
\usage{
quotations
}
\description{
A data set containing edited quotes from classic literature edited to be relevant for epidemiologists running into errors with their code.
}
\keyword{datasets}
