% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectquote.R
\name{selectquote}
\alias{selectquote}
\title{Select one random quote}
\usage{
selectquote(author = TRUE, book = TRUE, qpunct = "", nline = FALSE)
}
\arguments{
\item{author}{Logical element: TRUE (default) or FALSE. Indicates whether author name should be included with returned quotation.}

\item{book}{Logical element: TRUE (default) or FALSE. Indicates whether book title should be included with returned quotation.}

\item{qpunct}{Character element. The supplied character precedes and follows the quotation characters. Defaults to empty character. Suggest default, \code{"'"}, or \code{'"'}.}

\item{nline}{Logical element: TRUE or FALSE (default). Indicates whether newline characters (\code{'\\n'}) should be included where line breaks should be.}
}
\value{
Character element with randomly selected quote, followed by author and book (if indicated).
}
\description{
Select a quote from \code{quotations} data.frame (included in package) and return the quote along with the author and/or book (if indicated) as one character element.
Newlines, when included, are given as \code{'\\n'}.
}
\examples{
selectquote()
selectquote(book = FALSE)
selectquote(qpunct = '::')
selectquote(nline = TRUE)

}
\seealso{
\itemize{
\item \link[epierror]{selectquote.html} selects quotes with HTML formatting
\item \link[epierror]{getquotes} selects multiple quotes and return a data.frame
}
}
\author{
Philip Crain \href{mailto:philip.crain@doh.wa.gov}{philip.crain@doh.wa.gov}
}
