% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epierror.R
\docType{package}
\name{epierror}
\alias{epierror}
\alias{epierror-package}
\title{epierror: Quotes for the Epidemiolgist with an Error}
\description{
Easily include moody quotes in your error messages or emails.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DOH-PXC5303/epierror}
  \item Report bugs at \url{https://github.com/DOH-PXC5303/epierror/issues}
}

}
\author{
\strong{Maintainer}: Philip Crain \email{philip.crain@doh.wa.gov}

}
